Refactor a bit of systems
